+++ Compiler/Cli/CmdDoc.cpp
@@ -1,15 +1,14 @@
 // `rux doc`.
 
+#include "Cli/BuildReport.h"
 #include "Cli/Cli.h"
 #include "Cli/CompilerProgress.h"
 #include "Cli/DefineOption.h"
 #include "Cli/ManifestInput.h"
 #include "Cli/Reporter.h"
 #include "Documentation/Generator.h"
-#include "Driver/BuildReport.h"
 #include "Driver/BuildTarget.h"
 #include "Driver/CompilerDriver.h"
-#include "Formatter/Formatter.h"
 #include "Reporting/Reporting.h"
 #include "System/Os.h"